use orm for database, fix flow on register and login. #knownbug initial db not installed

+++ app/src/main/java/com/blogspot/ramannada/movieapp/SharedData.java
@@ -48,7 +48,7 @@ public class SharedData {
     }
 
     public void saveUserStorage(User user) {
-        editorSharedPreferences.putInt(USER_STORAGE_ID, user.getId());
+        editorSharedPreferences.putLong(USER_STORAGE_ID, user.getId());
         editorSharedPreferences.putString(USER_STORAGE_USERNAME, user.getUsername());
         editorSharedPreferences.putString(USER_STORAGE_EMAIL, user.getEmail());
         editorSharedPreferences.putString(USER_STORAGE_GENDER, user.getGender());
